Delete flaky NSUserDefaultsService spec
A single test in the suite was failing intermittently in test. I feel
comfortable deleting it without compromising confidence in the class
because there is another spec in the class's suite that both sets and
gets values through the service. That test has proved reliable, and
certainly covers the actual use cases that the production app will be
using the class for.