Add release notes draft
We don't want to have to remember every thing that goes into a new
release right when we create the release. What we'll do is maintain a
"draft" document between releases. Changes significant enough to
communicate in release notes should append into this draft document as
the changes get pushed. Then, when it comes time to make a release, we
move all the contents of the draft into a proper release notes
changelog.